32-bit support (#2115)
32-bit support

BRIG-878  #assign #start-work support for 32-bit #fixed #fixed

* eliminate the use of asLong and comparisons with LONG_MAX
   These are non-portable on 32-bit systems.

* miscellaneous fixes
add support to find validation scripts in developer environment when
the build directory is not under the source dir.

add missing NAME for logging test.

* update reference outputs
* update p4c refpoint

+++ bf-p4c/phv/phv.cpp
@@ -68,7 +68,7 @@ Container::Container(const char *name) {
     char *end = nullptr;
     auto v = strtol(n, &end, 10);
     index_ = v;
-    if (end == n || *end || index_ != v)
+    if (end == n || *end || index_ != static_cast<unsigned long>(v))
         BUG("Invalid register '%s'", name);
 }
